backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pypdf
pillow
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
